Exercise new taskq interface in splat-taskq tests
The splat-taskq test functions were slightly modified to exercise the new
taskq interface in addition to the old interface. If the old interface
passes each of it's tests, the new interface is exercised. Both sub tests
(old interface and new interface) must pass for each test as a whole to
pass.