Add Test: "Single task queue, recursive dispatch"
Added another splat taskq test to ensure tasks can be recursively
submitted to a single task queue without issues. When the
taskq_dispatch_prealloc() interface is introduced, this use case can
potentially cause a deadlock if a taskq_ent_t is dispatched while it's
tqent_list field is not empty. This _should_ never be a problem with the
existing taskq_dispatch() interface.

Signed-off-by: Prakash Surya <surya1@llnl.gov>