Add SPLAT test to exercise slab direct reclaim
This change adds a test case to the SPLAT suite to test the direct
memory reclaim functionality of the SPL slab implementation. Ideally,
all empty slabs should be able to be reclaimed quickly if the system is
under memory pressure. The new tests attempts to create these empty
slabs, create memory pressure, and then checks to ensure the slabs are
being reclaimed.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>