Append 'git describe' to SPL_META_RELEASE
Often times is it is useful to build packages from a non tagged commit,
and without editing the META file. In this use case, the 'Release'
assigned to the package is essentially meaningless. This commit fixes
the issue by appending information to the 'Release', which can then be
used to uniquely identify the commit the package was build from.

Signed-off-by: Prakash Surya <surya1@llnl.gov>