Detect kernels that honor gfp flags passed to vmalloc()
zfsonlinux/spl@2092cf68d89a51eb0d6193aeadabb579dfc4b4a0 used PF_MEMALLOC
to workaround a bug in the Linux kernel where allocations did not honor
the gfp flags passed to vmalloc(). Unfortunately, PF_MEMALLOC has the
side effect of permitting allocations to allocate pages outside of
ZONE_NORMAL. This has been observed to result in the depletion of
ZONE_DMA32 on Gentoo Linux. A kernel patch is available in the Gentoo
bug tracker for this issue:

https://bugs.gentoo.org/show_bug.cgi?id=416685

This negates any benefit PF_MEMALLOC provides, so we introduce an
autotools check to disable the use of PF_MEMALLOC on systems with
patched kernels.

Signed-off-by: Richard Yao <ryao@cs.stonybrook.edu>