Add 'Provides:' attributes for '-debug' packages
The generated '-debug' packages do not set their "Provides" attribute,
causing failed dependencies for packages which depend on "spl-modules"
or "spl-modules-devel". For example, with only the spl-modules-debug
package installed, installing the spl package will fail its dependency
check because it needs "spl-modules".

This patch fixes the problem by allowing the spl-modules-debug and
spl-modules-debug-devel packages to "Provide" spl-modules and
spl-modules-devel respectively.

Signed-off-by: Prakash Surya <surya1@llnl.gov>