SPLAT overcommit test
Reduce the object size in the slab overcommit regression test from
an order 6 to an order 1 allocation.  We still overcommit memory by
4x but the smaller object size reduces the odds of an OOM event due
to memory fragmentation.

This change was made to prevent this test case from triggering and
OOM which kills the buildbot test infrastructure.

In addition, move the kmem_cache_free() outside the spin lock.
Doing this under the spin lock isn't strictly safe, and if there
are large number of emergency objects allocated it be very slow.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>