Fix on_each_cpu autotools test
GCC 4.6.3 emits "error: called object '0u' is not a function" when
building the SPL against the Raspberry Pi Linux 3.2.27+ kernel on Gentoo
Linux. This causes the on_each_cpu() autotools check to fail in situations when it
should succeed.

I do not know why this issue does not occur on amd64, but we resolve
this issue by adding a dummy function.

Signed-off-by: Richard Yao <ryao@cs.stonybrook.edu>